requirements.txt
pinecone-client
streamlit-extras
streamlit==1.24.0
tiktoken==0.4.0
tqdm==4.65.0
requests==2.31.0
langchain==0.0.202
langchainplus-sdk==0.0.10
GitPython==3.1.31
youtube-transcript-api
pytube
transformers
torch
numpy==1.25.0
openai==1.3.2
openapi-schema-pydantic==1.2.4